refactor(agent-queue-worker): extract queue lifecycle from index.ts
Move worker events, queue depth polling, startup reconciliation,
shutdown handler, and signal handlers into queue/lifecycle.ts.

index.ts is now a ~55-line pure composition root.

Ref #1168
